AmbiguousDate: Make fmt mandatory
Since this class is used internally, the only benefit of having a
default value here is convenience in testing. On the other hand, it adds
an extra point of complexity in determining where default values come
from.

The default format should be set on a higher level that is exposed to
the user (e.g. command-line option).