Split strain and metadata outputs
Write the strain list directly instead of going through the metadata.
This is much faster on large datasets.

The side effect is that --output-strains is sorted alphabetically
instead of retaining the order from the original metadata. That order
was noted to be retained in 24.2.0 changelog but it's not explicitly
said anywhere else.