Allow spaces in discrete traits
Currently the parse fails on trees which contain discrete trains with spaces in them (for example, "South Africa"). This is because the regex patterns currently don't include a space in the list of valid characters. I simply added a space to the regex patterns for comments, strings within comments, and sets within comments. Though, I'm not sure if there are any side effects.